Criando camada completa de Service de Especialisa, criando ProfissionalService e definindo um service geral de cadastro

+++ gisa/src/main/java/com/fatec/gisa/entities/especialista/Especialidade.java
@@ -20,7 +20,7 @@ public class Especialidade {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idEspecialidade;
+    private Long idEspecialidade;
 
     private String nome;
     private String descricao;
