+++ gisa/src/main/java/com/fatec/gisa/repositories/especialista/EspecialidadeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.fatec.gisa.entities.especialista.Especialidade;
 
 @Repository
-public interface EspecialidadeRepository extends JpaRepository<Especialidade, Integer> {
+public interface EspecialidadeRepository extends JpaRepository<Especialidade, Long> {
 
     Optional<Especialidade> findByNome(String nome);
     boolean existsByNome(String nome);
